HW3
it was easy to get started as Mr Sami helpped me .

but due to my file moving . i deleted my HW by mistake 2 hours before my class for today at work and did not have the time to redo it.

I had to get it throw the code soltion.
I did with it step by step .